requirements.txt
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asciitree==0.3.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
boto3==1.40.16
botocore==1.40.16
branca==0.8.1
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.0
cycler==0.12.1
dask==2024.8.0
decorator==5.2.1
defusedxml==0.7.1
dicomweb-client==0.59.3
exceptiongroup==1.3.0
executing==2.2.0
fasteners==0.20
filelock==3.19.1
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2024.12.0
future==1.0.0
GDAL==3.12.0.1
geojson==3.2.0
Glymur==0.13.8
h5py==3.14.0
hdbscan==0.8.40
hf-xet==1.1.8
-e git+https://github.com/HistAI/hibou.git@c453bbe4dab0fec6f7df343b09ea87048629c58d#egg=hibou
huggingface-hub==0.34.4
idna==3.10
imagecodecs==2024.12.30
imagecodecs-numcodecs==2024.12.30
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipyleaflet==0.20.0
ipython==8.18.1
ipyvue==1.11.2
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter-leaflet==0.20.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
large-image==1.32.11
large-image-converter==1.32.11
large-image-source-bioformats==1.32.11
large-image-source-deepzoom==1.32.11
large-image-source-dicom==1.32.11
large-image-source-dummy==1.32.11
large-image-source-gdal==1.32.11
large-image-source-mapnik==1.32.11
large-image-source-multi==1.32.11
large-image-source-nd2==1.32.11
large-image-source-ometiff==1.32.11
large-image-source-openjpeg==1.32.11
large-image-source-openslide==1.32.11
large-image-source-pil==1.32.11
large-image-source-rasterio==1.32.11
large-image-source-test==1.32.11
large-image-source-tiff==1.32.11
large-image-source-tifffile==1.32.11
large-image-source-vips==1.32.11
large-image-source-zarr==1.32.11
llvmlite==0.43.0
locket==1.0.0
lxml==6.0.1
mapnik==4.1.1.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
morecantile==6.2.0
mpmath==1.3.0
multidict==6.6.4
nd2==0.10.3
networkx==3.2.1
numba==0.60.0
numcodecs==0.12.1
numpy==1.26.4
ome-types==0.6.1
opencv-python==4.5.5.64
openslide-bin==4.0.0.8
openslide-python==1.4.2
packaging==25.0
palettable==3.3.3
pandas==2.3.2
parso==0.8.5
partd==1.4.2
pexpect==4.9.0
pillow==11.3.0
pillow-jpls==1.3.2
pillow-jxl-plugin==1.3.4
pillow_heif==1.1.0
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
pydicom==2.4.4
Pygments==2.19.2
pylibjpeg-openjpeg==2.3.0
pylibmc==1.6.3.1
pylibtiff==0.6.1.1
pynndescent==0.5.13
pyparsing==3.2.3
pyproj==3.6.1
python-bioformats==4.1.0.8.3.0
python-dateutil==2.9.0.post0
python-javabridge==4.0.4.1
pytz==2025.2
pyvips==3.0.0.1
PyWavelets==1.6.0
PyYAML==6.0.2
rasterio==1.4.3
rawpy==0.25.1
redis==6.4.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
resource-backed-dask-array==0.1.0
retrying==1.4.2
rio-cogeo==5.4.2
rpds-py==0.27.0
s3transfer==0.13.1
safetensors==0.6.2
scikit-image==0.19.3
scikit-learn==1.6.1
scipy==1.13.1
shapely==2.0.7
simplejpeg==1.8.2
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tifftools==1.6.1
tokenizers==0.21.4
tol_colors==2.1.0
toolz==1.0.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
transformers==4.55.4
triton==2.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
universal_pathlib==0.2.6
urllib3==1.26.20
wcwidth==0.2.13
widgetsnbextension==4.0.14
wsidicom==0.20.6
xformers==0.0.27
xsdata==24.3.1
xyzservices==2025.4.0
yarl==1.20.1
zarr==2.18.2
zipp==3.23.0
